fixing migration dropping storytime table

+++ db/migrate/20150905173353_drop_storytime_comments_table_from_db.rb
@@ -1,6 +1,8 @@
 class DropStorytimeCommentsTableFromDb < ActiveRecord::Migration
   def up
-    drop_table :storytime_comments
+    if ActiveRecord::Base.connection.table_exists? :storytime_comments
+      drop_table :storytime_comments
+    end
   end
 
   def down
